+++ laktory/models/datasinks/tabledatasink.py
@@ -54,6 +54,7 @@ class TableDataSink(BaseDataSink):
         name = self.table_name
         if name is None:
             return
+
         # names = re.findall(r"\$\{vars\.[^}]+\}|[^.]+", name)
         names = re.split(r"\.(?![^{}]*})", name)